Fix the issue that the overriding AE_MODE doesn't work correctly for the triggerAf request
When triggerAf is invoked, it should override the AE_MODE to AE_MODE_ON regardless of
the current flash mode to avoid flash being triggered during tap-to-focus
on some devices.

But there is a regression that causes it failed to override.

Solution: Add a new OptionPriority - HighPriority and use
that priority to set the AE_MODE option for the workaround.

Will have a follow-up CL to add the testing.

Relnoe: "Fixed the issue where flash is triggered during
tap-to-focus in low lite on some devices."
Test: manual
Bug: 341029779

Change-Id: I06d003756da133388f525766c4bfce57c09e4bd0