Release layers when the node is not placed anymore
Previously I was releasing the layers when the LayoutNode is deactivated, however it is not a proper lifecycle callback. We "attach" the layer to the LayoutNode when placeWithLayer() is called. And we should "detach" when the LayoutNode is not placed anymore.

This change is also a prerequisite for enabling layers reuse as if we don't "detach" layers properly from the LayoutNode, it will not be possible for us to handle the case when the layer is released. then it is recycled and the same placeable is placed with the same layer object, however the recycler layer is not considered the same anymore and should be "re-attached".

Bug: 340651308
Test: new test in DrawingPrebuiltGraphicsLayerTest
Change-Id: I8e904e58b183030501abc14efef319a275be96e3