webkit: mark non-enforced methods as @UiThread.
These methods must all be called on the UI thread for correct behavior
(the underlying implementations are not threadsafe), but unfortunately
this was not documented or enforced when calling the API.

Add the @UiThread annotations as the first step toward fixing this, but
don't actually introduce any thread checks yet.

Bug: 342204117
Test: n/a
Relnote: Methods are now documented with @UiThread where appropriate.
Change-Id: I6c7e092a07cd47c4b76f5cffd2628eda3127edd9