revert core-telecom adapter changes
Some crash reports came in with a stack trace that the telecom package
could not be found when registering a phone account.  The recent
adapter changes look to have caused this. One other reason for the
crash is that the client app is using too loose of telecom support
check.

Bug: 342834226
Bug: 342902827
Test: existing test
Change-Id: I8f4b340b5978958bb16e62e50f301215c70b081f