Bugfix: In Preview, invalidate camera edge when the app Surface is invalidated
The bug happens when using effect targeting only Preview with PreviewView in PERFORMANCE mode.

Currently, we send the existing SurfaceRequest again, when the app edge is invalidated. It's no longer working since aosp/2699899. Now that the child DeferrableSurface is closed when parent is closed, we will not be able to propagate a new Surface.

In this CL, we also invalidate the parent, which makes it connectable again. This adds a bit of churn when SurfaceView is used and the preview lifecycle ends, but it's safer.

Bug: 343710980
Test: manual test and ./gradlew bOS
Change-Id: I6d9cdfcbd47f646aa3d9d9d2ee79dd28b31cef6a