Decouple ScrollableStates from Navigator
Previously, the Navigator kept two parallel lists, one for the menues
in the navigation path and one for the ScrollableStates on them.
The Saver implementation didn't save and hence couldn't properly
restore the list of ScrollableStates, so when the apps state changes
and it's restarted (in a phone this can easily be tirggered by rotating
the phone, not sure on wear) there is a mismatch and then next time
there is a back navigation the app crashes.
This change removes the responsavility of keeping the ScrollableStates
from the Navigator and instead each Demo that needs it keeps, reusing
it or creating it as needing.

RelNote: "Fixed a possible crash when an app state change forces a restart"
Test: Manual
Change-Id: I7b7976d5ce9dd605c9dd0fd24aa0698a556853b3