Resolve invalid code generation in iOS.
When a Map with a value of type List was encountered in Kotlin codegen,
it was incorrectly getting initialized as a new ArrayList instance. The
same issue was discovered when using Set value types as well. This CL fixes
both issues by utilizing the appropriate `mutableListOf()` and
`mutableSetOf()` functions in code generation.

Bug: 340983093
Test: BaseQueryTest.kt
Change-Id: If1404d0b3a5da5a2fd03dd89c3dc282f9b54138f