Fix a case where movable content was losing invalidations
When movable content is in a sub-composition (such as BoxWithCostraints
or LazyColumn) and it observes state that has changed but the change
also causes it to move out of the sub-composition these invalidations
can be lost.

This change now ensures that the invalidations of sub-composition that
contains moving content will be correctly recorded in the moving
content.

Fixes: 343178423
Test: new test, ./gradlew :compose:r:r:tDUT
Change-Id: I24c5915cfe034fe74a70986cd5cce22b278e71f4