Optimize Matrix computations
The implementation of graphics layer needs to apply a TRS matrix
(translation/rotation/scale) which was previously done by computing
separate transforms successively. This was not only inefficient
from a CPU perspective, it also led to the allocation of 3 temporary
matrices. This was the cause of thousands/dozens of thousands of
allocations when scrolling lazy containers.

This CL combines all the operations necessary to create a TRS
matrix around a pivot point into a single method called
resetToPivotedTransform().

With this change, updating a layer's matrix goes from
1062 instructions and 39 branches down to 168 instructions and
1 branch.

More importantly, all allocations are eliminated.

Another important optimization is the addition of array size
checks all over the file. By construction we know the array
always has a size of 16, but the runtime cannot guarantee
this and inserts a lot of bounds checks. This is especially
bad when we do multiple accesses as it generates a lot
of instructions and branches we do not need.
By adding size checks manually, we give a hint to ART that
the check should happen only once, greatly reducing the
number of extra instructions. For instance in the
isIdentity() method, this removes half of the instructions
and half of the branches.

Additional optimizations:
- Implement reset as a series of assignments. Saves 9
  instructions and 6 branches
- Do the same in setFrom, saving 7 branches for a few
  more instructions
- Inline dot(), making timesAssign go from 201 instructions
  and 7 branches down to 189 and 4
- Rewrite isIdentity without a loop. For matrices that
  benefit from early returns, it runs 10% faster. For
  identity matrices, it runs 80% faster

Test: MatrixTest, compose-ui tests
RelNote: N/A
Change-Id: Ia478735b8ddcc182ac1aaf82d013e4e14832a9c5