Propagate errors correctly and retry on camera-specific errors
We currently only restart `Camera2CameraController` when the last error
is `ERROR_CAMERA_DEVICE`. This is unnecessarily stringent. When a camera
becomes available (`onCameraAvailable`), it means it is available to be
used and we should be able to retry unless we have a graph configuration
error, which would suggest issues with the `CameraGraph.Config` and that
a retry would not succeed.

This is a re-land of aosp/3098281 with the quirk for `onCameraAvailable`
removed.

Bug: 340583180
Test: CameraPipe and CameraX test app in multi-resume mode.

Change-Id: I8c61444777517a0f5e7ee064a7e38899d6fc940a