Avoid API 26/27 crash by not traversing assist structure
SDK_INT O and SDK_INT O_MR1 are missing a check to see if a virtualView is pointing to its host. Without this check, populating ANIs will become stuck in a loop traversing between parent and child indefinitely, resulting in the stack overflow. To avoid this crash, on API 26 and 27, Compose will not send the assist structure when services like Voice Assistant request for it. This does not affect TalkBack or Autofill.

Test: Manually verified that Voice Assistant does not crash before and after change.
Bug: 331271890
Bug: 251152083
Change-Id: I0559eb004e83bc3152e871ff3450b16a3756bdb4