Add index-based operations to DraggableAnchors
This is based on API council feedback.
It allows us to move forEach out of
the DraggableAnchors interface and
make it an inline extension instead.

Bug: 344652977
Relnote: Renamed methods on DraggableAnchors. `hasAnchorFor` is now called `hasPositionFor`, `minAnchor`/`maxAnchor` are called `min/maxPosition`, and introduced two new methods `anchorAt` and `positionAt` that allow iteration. `DraggableAnchors#forEach` is now an inline extension method.
Test: Existing

Change-Id: I824dab728676064a7f9c334a940618f5597a4a71