ICS integration of app actions - CallExtensions
Part 1: Re-define CallExtensions to not use
InCallServiceCompat directly

Originally, a tight coupling between the creation
of extensions and onCallAdded was done in order to
try to "hide" the complexity of creating a CallCompat
and only providing a signal to the user when cap exchange
was complete. This was not possible to do, however, because
the app needs to provide their capabilities before
capability exchange starts on a call-by-call basis anyway.

Instead, we are decoupling the InCallServiceCompat from
the capability exchange code in order to create a more
composable, decoupled approach to defining/maintaining
extensions. As with CallsManagerExtensions, the CallExtensions define
the class CallExtensionScope, which allows for the initialization
and usage of extensions in the following way:

connectExtensions(context, call) {
  // initialize extensions using registerExtension internally
  onConnected {
    // Cap exchange is complete for extensions and the extensions
    can now be used
  }

This is part 1 in a larger refactor. In this CL, the CallExtensions
class was defined with a method to register extension for capability
exchange, which will replace the existing implementation
in follow up CLs. In the next parts, the participant extension will
be defined and then the obsolete implementation will be removed in
favor of the new implementation. Finally, we will move the impl into
APIs.

Bug: 338651142
Test: core-telecom tests
Change-Id: I9592afdfd0aad8fffee60f70bb62ac056d699941