Fix running tests via IDEA configuration desktop/test.run.xml (#1386)
Tests don't start after the first failure at the moment.

Note, that this option makes the Gradle task itself always successful,
even if there are failed tests, so it shouldn't be used in Terminal. It
is okay if we run it from IDEA - it doesn't look at the task result.