Add Annotation type for styles/annotations and map functions to AnnotatedString
This change introduces a new marker type Annotation for annotations (including SpanStyle and ParagraphStyle) in the AnnotatedString. It allows us to easily store them all together in a single list. All this enables a few capabilities
* we provide mapAnnotation and flatMapAnnotations functions that allows to support links theming in material; it also enables possibilities for html/markdown support
* by preserving an order between styling and annotations, we allow links for example to participate in nested styling for cases like `withStyle(A) { withLink(B) { withStyle(C) } }`; it also enables similar capabilities for custom annotations which opens more possibilities for html/markdown support

Test: compose:ui:ui-text:cC, compose:ui:ui-text:test, BasicTextLinkTest

Bug: 329052426
Fixes: 317894981
Relnote: "`AnnotatedString`'s annotations (icluding `SpanStyle` and `ParagraphStyle`) now all implement the marker interface `Annotation`. The list of annotations is made public, so it can be accessed directly as well as set directly during the `AnnotatedString` construction. We also introduced two methods `mapAnnotations(transform)` and `flatMapAnnotations(transform)` that create a new `AnnotationString` with annotations consisting of the results of applying `tranform` function to each annotation in the original list of annotations"

Change-Id: I7d542d90308a6b1d439371ced43c4f63deb4575f