Ensure that z-order considers predictive back
Currently, when doing a predictive back gesture with Navigation Compose,
the z-order does not consider the predictive back as a pop and as a
result the incoming screen is on top of the screen being popped. This is
incorrect.

We should ensure we consider predictive back for z-order the same way we
do for the enter and exit transitions.

RelNote: "When using Predictive back with Navigation Compose, the
destination being popped will now correctly animate on top of the incoming destination."
Test: added ScreenShotTest
Bug: 345993681

Change-Id: I2077b399fc9bc9a846da6ed17d63d002c8e8122f