Adding a new Material3 loading indicator component
- Introducing new LoadingIndicator APIs.
- Adding a loading indicator option to the PullToRefreshBox.
- Updating the pull refresh tokens to use the LoadingIndicator color tokens.
- Adding samples and tests.

Test: LoadingIndicatorBenchmark, LoadingIndicatorTest, LoadingIndicatorScreenshotTest
Bug: 346639009
Relnote: "Introducing a new Composables for Material 3
LoadingIndicators. The API supports a determinate and an indeterminate
variations of the indicator. The determinate one morphs, by default,
between two shapes with a given progress value, while the indeterminate
one animates and morphs continuously between several predefined Material
shapes (RoundedPolygons).
You may customize the type of shapes and the number of shapes when
setting up a LoadingIndicator as you like."

Change-Id: Ie199ec68c5eae77629bbcfc319183b3d104e3223