Optimizing for Dogfood Across Beta Versions
Given bugfixes occured, certain beta versions need to be supported so
that dogfooding can go smoothly. We made decisions to modify how the
framework bundled biometric prompt data types (and the naming of certain
constants), but decided we still want
to allow dogfooders in beta to access these changes.

Thus, while changes were made with the 'stable' release in mind, this
change allows beta users to continue conditionally testing our framework
code, given the framework change in question was a re-organization of a
unit of code (namely migrating to a bundle from directly imbuing
properties into the slice and changing constants).

Bug: 342447585
Test: Device Testing across Beta versions, unit tests
Change-Id: I6b785609af675708bb7523eb35dcdf8bf7fca651