Fix StackOverflow when using navigation with type safe setPopUpTo (#1410)
The issue is that the call to setPopUpTo becomes recursive since the
`hashCode` is treated as `route: T`. The fix is to duplicate what the
other functions are doing, which isn't too bad because there isn't a lot
happening.

Another option would be to extract everything into a separate private
function and have all the overloads call that, but it seemed like
overkill here.

## Testing
Describe how you tested your changes. If possible and needed:
- Test it on a sample project

## Google CLA
Signed

---------

Co-authored-by: Ivan Matkov <matkov.ivan@gmail.com>