Move `text:text` package into `compose:ui:ui-text`
`text:text` package was originally designed to remain compose free but be a utility provider for the Compose ui-text package. However, the way it's set up prevents IDE from correctly indexing classes and makes it difficult to work in foundational classes inside `ui-text`. The module is not used anywhere else and only included as a source set.

Test: gradle :compose:ui:ui-text:cAT
Change-Id: Ifbfdc29d02390a8073b13e61b46c0a0b76acd469