Increase max buffer for persistence ImageReader
Because the internal ImageReader previously had a max
buffer size of 1, both acquireNextImage and acquireLatestImage
have the same behavior. However, the implementation of
acquireLatestImage calls acquireNextImage until it returns
null, after the first call, the max number of buffers has
already been acquired (1) and therefore the subsequent call
will log a warning.

Additionally previously using a max buffer size of 1 means
that there could be the potential for the persistence render
to block until the previous persistence render completed.
So instead increase the max number of buffers to 3 as that
is the default depth for swapchain. Buffers will not be allocated
unless one is necessary, so if the producing/consuming sides
are fast enough, only 1 buffer should be allocated and reused
internally anyway.

Fixes: 343938070
Test: re-ran compose tests + verified with sample application
Change-Id: I44864b9b570b6df1d238a1c671b3a37299ef34ce