Avoid API 23-25 crash by not populating assist structure
API levels 23 through 25 are missing a check to verify an ANI is not null before trying to populate it. Without this check, population APIs can potentially be called on null object references. To avoid this crash, on API 23 through 25 (in addition to 26 and 27), Compose will only minimally populate the assist structure. This does not affect TalkBack or Autofill.

Test: Manually verified that Voice Assistant does not crash before and after change.
Fixes: 320768586
Change-Id: Ia87217de73b619feea0c53e4ffbd86d914ff5a44