Fix bug with selecting same complication in editor
After editing a complication, we clear the slot to avoid seeing a
glimpse of the previous complication. Normally an updated
complication is quickly sent, however if the user selected the
same complication this doesn't happen.

This patch modiifes the logic so we don't clear the slot if the
data source didn't change.

Bug: 349494503
Test: I was able to reproduce the issue in the video with an androidx sample and I can verify this patch fixes it.
Change-Id: I8833fb391c7b1578b0d28391a82b2afdee02344c