Fix persistence render support with SurfaceView
Update layer persistence logic to handle persistence
renders to into ImageReader instances that are owned
by each layer instead of leveraging a shared ImageReader.
This allows for a persistence render to be done once for
a layer upon creation instead of requiring a persistence
render for all layers whenever a new layer is introduced.

Fixes: 347782855
Test: Manual + re-ran compose tests
Change-Id: I00bcf583dfe6efb3b60f5605821a768f4034fea6