Fix Abstract method error.
We are seeing an Abstract method error when using the functional
interfaces in extensions.core. This was introduced when we switched to
compile only. The fix is to provide duplicate interfaces and have added
proguard rules to keep the duplicate interfaces.

Bug: 350514774
Test: minify the samples and run through regular flows.
Change-Id: I1072da0f3d1ebc398e4ece49d3739f1e0355b082
Merged-In: I1072da0f3d1ebc398e4ece49d3739f1e0355b082