Attach scroll changed listener based on window attachment/detachment
Couples the lifecycle of the scroll change listener added inside
SandboxedSdkView to SandboxedSdkView's window attachment. Previously,
the listener was added based on surface changed callbacks, which
caused issues in backwards compatibility mode where no
additional surfaces are used.

In a future change we should be able to move other callbacks to
be aligned with the SandboxedSdkView's window lifecycle.

Test: manual test app viewability layer in backwards compat mode
Bug: 339377737
Change-Id: Ibfffad1d4533feca7ae5a34f683bd7381622480d