Improve behavior and test coverage  of CameraGraph.useSessionIn
The primary goal of this CL is to address delay() and exception
handling when using a session. Without this change, running
tests inside `runTest` or `runBlocking` with useSessionIn can
cause the test method to timeout or stall. This change also
includes a test to ensure this behavior functions correctly:

```
@Test
fun useSessionInWithRunBlockingDoesNotStall() = runBlocking {
    val deferred = cameraGraph.useSessionIn(this) { delay(1) }
    deferred.await() // Make sure this does not block.
}
```

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: I11fa030607ec27e116947724e562deaed04f6398