[Dual Camera v2] Fix the race condition when switching between single and dual cameras
Sometimes the single camera preview doesn't work correctly when switching between single and dual cameras.

The root cause is LifecycleCamera for dual camera is being reused for single camera case because the id of CameraUseCaseAdapter is not generated correctly. So the UseCaseActive is triggered for secondary camera in single camera case.

Bug: b/351049459
Test: test on real devices
Change-Id: Ifeb5b87bc434da4467821c13c5b949120b417eb1