Resume / Stop Player view based on PlayerView visibility
To determine video ad player visibility, we look for value of onScreenGeometry.
Ad is considered to be not visible if host screen is in stopped state.
Ad is in scroll view and user scrolled the host such that ad is no more visible,
Visibility in such cases is determined by value of onScreenGeometry
received in onUiContainerChanged callback function.
onScreenGeometry will have empty rectange bounds if ad is not visible

Bug: 347679917
Test: Manual
Change-Id: If70374f1abbd35ecad474e29aa40e7ff4821ab43