add a placeholder value in the acct extras; bump alpha 9
If a telecom account contains an empty bundle and the device is
rebooted, the account extras will be null upon restart.  However, if a
placeholder value is added to the extras, the bundle will be non-null on
restart.  In oder to prevent crashes on OEM devices that expect non-null
extas, core-telecom will add a placeholder value to the account extras.

Fixes: 352526256
Relnote: add a placeholder value to extras bundle
Test: manual + unit test
Change-Id: Iebf7fdfb5bab35ec429dabb093d351868ca3edcc