Ensure onBackInstance is updated on PredictiveBack
There is an issue where if you do a system back with PredictiveBack
enabled, the PredictiveBackHandler never clears the
`onBackInstance`. So if you follow the system back with a regular
`onBackPressed()` on your `OnBackPressedDispatcher`, the back pressed is
ignored.

We need to ensure that after we handle a Predictive back gesture and it
either completes or cancels, we change the isPredictive flag of the
`onBackInstance`. We can't clear it because of the interrupt case, but
changing the `isPredictive` flag to false is enough.

RelNote: "Fixed an issue with `PredictiveBackHandler` where after doing
system back, attempting to do `onBackPressed` from the
`OnBackPressedDispatcher` would fail."
Test: PredictiveBackHandlerTest
Bug: 351083458

Change-Id: I5f1f8075fb3abac768761e0e8889254d4c31f81e