Do not throw ObjectUnavailableException when createHighSpeedRequestList
Currently,
AndroidCameraConstrainedHighSpeedCaptureSession.createHighSpeedRequestList
throws ObjectUnavailableException when an exception is thrown by the
framework. This is currently not caught inside CameraPipe, but thrown
inside CameraPipe's own CoroutineScope, which means it's uncatchable for
users. We handle this by instead returning a null List<CaptureRequest>
and mark it as a request submission failure.

Bug: 332705637
Test: CameraPipe test app.

Change-Id: Icbcc145390b7fbf5458ab4eed25fc20b809817d5