Remove the custom A11Y actions from the DatePicker
Remove the custom actions we had for the years selection grid in the
Material3 DatePicker.
The custom actions were intended to provide navigational action to
scroll the years list up and down, instead of having the list
auto-scroll when TalkBack is on the the user scrolls horizontally. This
ended up causing more confusion, so we are removing them.

Bug: 350404866
Bug: 266613839
Bug: 266610993
Test: Manual with TalkBack on

Change-Id: I3f0209e2bce37f3de9ce65960f8d8f2f3055cdb9