waitForIdle() now waits for drawing to complete
Fixes: 355076437

ComposeTestRule.waitForIdle() previously waited for all compositions,
but did not wait for drawing to complete. This means that if
runOnIdle() was used, we could sometimes enter before the draw()
and most of the times after the draw(). This CL forces waitForIdle()
to wait for after the frame draw before continuing.

Test: ran flaky test 800 times without failure
Change-Id: I463567267ecf6e3edc75de09da141eaca007a09c