Add performFirstLinkClick link testing API
In this change we started publishing the semantics of the links. Because links don't have Link roles, to identify them we use a new semantics property LinkTestMarker. Note that Text node might have a clickable child that could either be a link or an inline content.
At the same time on Android the links are being made accessible through the UrlSpan/ClickableSpan therefore we exclude them from the currentNodes and hit testing in the ACVADC.

Relnote: "Add performFirstLinkClick API that allows to click on a link/clickable portion of the text for testing purposes"
Bug: 351885689
Test: LinkClickTest
Test: links demo with Talkback, SA and SelectToSpeak on
Change-Id: Icd0becab6818712ab7fa6f3abe83f441a841af5e