Add OrderedScatterSet
This new collection is an ordered variant of ScatterSet that preserves
the insertion order on iteration. This can be useful to track which
items were added first/last. It is built on the same principles as
the Scatter family of collections, providing an allocation-free usage
unless resizing is required.

Test: OrderedScatterSet, ScatterSet
Relnote: OrderedScatterSet is a new ordered, allocation-free collection
Change-Id: Ic417847247396583835d88abfb05b0af25224435