Streamline set parameter flow: direct call setParameterAsync on RequestControl
This change simplifies the process of setting parameters asynchronously by
removing an unnecessary layer of indirection.
Previously, the setParameterAsync function was part of the UseCaseCamera interface,
which then forwarded the call to RequestControl.

To improve code clarity and efficiency, this change allows direct invocation of
setParameterAsync on the RequestControl object. This eliminates the extra step
through UseCaseCamera, making the code more straightforward.

Bug: 358338684
Test: Robolectric test and device test
Change-Id: I2a68f78c84d2cb44f485dfe823cbd0e656169970