Prevent setting seekable transition playTime when no transition is running
`seekAnimations(playTimeNanos = 0)` is called after animations finish. It's triggered by the animation duration being updated to 0.
It seems reasonable for a transition that's not playing (`startTimeNanos == UnspecifiedTime`) to _stay_ not playing when seeked with a 0-duration animation, and `playTimeNanos = 0` ambiguously means 0% or 100% when the duration is 0. Setting that to 0 when there's no animation erroneously causes `isRunning` to always return true. This defends against that by checking for 0 duration and preventing calling seekToFraction when that's the case.

Fixes: 334275648
Test: Added test asserting parent seekable and child transition return correct isRunning before, during, and after animateTo
Change-Id: I227113399a9b6baaa3496eaa0a3661a7a11bebb3