Fixed the crash when bindToLifecycle is invoked with destroyed lifecycleOwner
A crash might happen if bindToLifecycle is invoked with the destroyed
lifecycleOwner. Although this can be avoided by checking
the lifecycle status before calling bindToLifecycle, it is a common
mistake apps often make and since it happens in edge cases, most apps
are not aware of this issue until user reports the crash in analytics.

Solution: return an inactive camera instance even when lifecyeOwner
is destroyed.

Bug: 359114636
Relnote: "Fixed the crash when bindToLifecycle is invoked with
destroyed lifecycleOwner."
Test: ProcessCameraProviderTest

Change-Id: I6e6d8f0c545730987587c318bae92998c06f6947