Add LevelIndicator for Material3 for indicating the value of settings such as volume.
The LevelIndicator is similar to the ScrollIndicator, but is displayed on the opposite side and has different stroke width and color.

BUG: 334026179
Test: Added unit tests and screenshots
Relnote: "We have added LevelIndicator to our Material3 library of components. LevelIndicator is used to show the value of a setting such as volume, and can be used with the existing Stepper component to construct a volume screen. LevelIndicator is similar to ScrollIndicator, but is displayed on the opposite side of the screen and has a wider stroke width and different indicator color by default."
Change-Id: I8a4ac3d8ee7e06a1d73642aeee731ecc29298220